requirements.txt
selenium_stealth
selenium
chromedriver-py
webdriver-manager
Pillow
types-requests
